Rename package annotations -> annotation
It looks like it is common to name the annotations package 'annotation', so let's keep that convention
