+++ contentgrid-spring-data-rest/src/testFixtures/java/com/contentgrid/spring/test/fixture/invoicing/model/Order.java
@@ -1,6 +1,6 @@
 package com.contentgrid.spring.test.fixture.invoicing.model;
 
-import com.contentgrid.spring.querydsl.annotations.CollectionFilterParam;
+import com.contentgrid.spring.querydsl.annotation.CollectionFilterParam;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import java.util.HashSet;
