+++ contentgrid-spring-querydsl/src/main/java/com/contentgrid/spring/querydsl/annotation/QuerydslPredicateFactory.java
@@ -1,4 +1,4 @@
-package com.contentgrid.spring.querydsl.annotations;
+package com.contentgrid.spring.querydsl.annotation;
 
 import com.querydsl.core.types.Path;
 import com.querydsl.core.types.Predicate;
